footer and alert added

+++ src/PrivateRoute/MyServices.jsx
@@ -3,6 +3,7 @@ import { AuthContext } from "../AuthProvider/AuthProvider";
 import useAxiosSecure from "../Hooks/useAxiosSecure";
 import TableRow from "./TableRow";
 import { Helmet } from "react-helmet-async";
+import Swal from "sweetalert2";
 
 const MyServices = () => {
     const { user } = useContext(AuthContext);
@@ -19,15 +20,33 @@ const MyServices = () => {
     }, [axiosSeceure, url]);
 
     const handeleDelete = _id => {
-        fetch(`https://assignment-11-server-one-sandy.vercel.app/services/${_id}`, {
-            method: "DELETE"
-        })
-            .then(res => res.json())
-            .then(data => {
-                const remaining = myServices?.filter(prods => prods._id !== _id);
-                setMyServices(remaining)
-                console.log(data)
-            })
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`https://assignment-11-server-one-sandy.vercel.app/services/${_id}`, {
+                    method: "DELETE"
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        const remaining = myServices?.filter(prods => prods._id !== _id);
+                        setMyServices(remaining)
+                        console.log(data)
+                    })
+                Swal.fire({
+                    title: "Deleted!",
+                    text: "Your file has been deleted.",
+                    icon: "success"
+                });
+            }
+        });
+
     }
 
     return (
